changed readme and added bash file
