+++ package.json
@@ -9,10 +9,14 @@
     "cypress:open": "cypress open",
     "lint": "npx eslint --ext .ts,.js .",
     "on-prem": "npm run cypress:run",
-    "on-prem-suite": "npm run cypress:run",
-    "docker": "docker-compose up",
+    "on-prem-default" : "npm run cypress:run 'cypress/integration/e2e/endtoend.spec.ts'",
+    "on-prem-suite": "npm run cypress:run", 
+    "docker": "npm run clean:reports && docker-compose up && docker-compose down",
+    "docker-default": "npm run clean:reports && spec='cypress/integration/e2e/endtoend.spec.ts' docker-compose up && docker-compose down",
     "bstack-single": "browserstack-cypress --config-file run_conf/bstack-single.json run --sync --spec",
+    "bstack-single-default": "browserstack-cypress --config-file run_conf/bstack-single.json run --sync --spec 'cypress/integration/e2e/endtoend.spec.ts'",
     "bstack-local" : "browserstack-cypress --config-file run_conf/bstack-local-single.json run --sync --spec",
+    "bstack-local-default" : "browserstack-cypress --config-file run_conf/bstack-local-single.json run --sync --spec 'cypress/integration/e2e/endtoend.spec.ts'",
     "bstack-local-parallel": "browserstack-cypress --config-file run_conf/bstack-local-single.json run --sync --parallels",
     "bstack-local-parallel-browsers": "browserstack-cypress --config-file run_conf/bstack-local-parallel.json run --sync --parallels",
     "bstack-parallel": "browserstack-cypress --config-file run_conf/bstack-single.json run --sync --parallels",
