+++ run_all_profiles.sh
@@ -1,0 +1,10 @@
+npm run on-prem-default
+npm run docker-default
+npm run bstack-single-default
+npm run bstack-parallel
+npm run bstack-parallel-browsers
+./BrowserStackLocal --key JNzsLXukmxpiaaCsBhSx --daemon start
+npm run bstack-local-default
+npm run bstack-local-parallel
+npm run bstack-local-parallel-browsers
+./BrowserStackLocal --key JNzsLXukmxpiaaCsBhSx --daemon stop
